# app/Models/Order.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Order extends Model
{
    protected $fillable = [
        'user_id',
        'total', 
        'comment',
    ];
    
    public function items()
    {
        return $this->hasMany(OrderItem::class);
    }
}